Add "default project" option to make annotations easier for common case
This causes ambiguous service account names with no host-part
(eg. "foo", not "foo@PROJECT.iam.gserviceaccount.com") to be automatically
treated as a user-managed account for the given default project.

This is a huge usability gain in the common case,
as it means most users can simply annotate as "foo", instead of "foo@my-project.iam.gserviceaccount.com",
which is especially annoying when the project id may be generated elsewhere.